change N from valarray<size_t> to <int> (does not make sense but requested by Sachiko)

+++ spk/cspk/ml/GridIntegral.h
@@ -8,7 +8,7 @@ void GridIntegral(
 	double (*F)(double *X, size_t m, void *p)  ,
 	size_t                                 m   ,
 	void                                  *p   ,
-	const std::valarray<size_t>           &N   ,
+	const std::valarray<int>              &N   ,
 	const std::valarray<double>           &L   ,
 	const std::valarray<double>           &U   ,
 	double                                &integralEstimate,
